requirements.txt
line-bot-sdk
flask
configparser
requests
python-dotenv
uwsgi